+++ src/Faker.jl
@@ -7,6 +7,8 @@ import YAML
 data=Dict()
 data["faker"]=Dict()
 
+tostring(a::String,b::String)::String = string(a," ",b)
+
 function language(file::String) #"es-MX"
 	original= pwd()
 	cd(abspath(joinpath(dirname(Base.find_package("Faker")), "../src/locales")))